% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_grid_builder.R
\name{bi_find_steps}
\alias{bi_find_steps}
\title{bi Steps Builder}
\usage{
bi_find_steps(lb = -Inf, rb = Inf, a, th, mode, mode_i, mode_n, f)
}
\arguments{
\item{lb}{scaler density lower bound}

\item{rb}{scaler density upper bound}

\item{a}{scaler step area}

\item{th}{scaler pre_acceptance threshold}

\item{mode}{vector density modes}

\item{mode_i}{mode index}

\item{mode_n}{total number of modes}

\item{f}{density function}

\item{h}{log transform of the density function}

\item{h_prime}{first derivative of h}
}
\description{
bi Steps Builder
}
