% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delete_build_in_grid}
\alias{delete_build_in_grid}
\title{Delete Built-in Grids}
\usage{
delete_build_in_grid(sampling_function, grid_type = "custom")
}
\arguments{
\item{sampling_function}{String. The name of the sampling distribution's function in STORS.
For example, \code{"srgamma"} or \code{"srchisq"}.}

\item{grid_type}{String. Either \code{"custom"} to delete the custom grid or \code{"scaled"} to delete the scaled grid.
Defaults to \code{"custom"}.}
}
\value{
A message indicating the status of the deletion process, or an error if the operation fails.
}
\description{
This function deletes built-in grids from disk by specifying the sampling function and grid type.
It is useful for managing cached grids and freeing up storage space.
}
\details{
The function looks for the specified grid type associated with the sampling function in the built-in grids directory.
If the grid exists, it deletes the corresponding grid file from disk and frees its cached resources.
If the specified sampling function or grid type does not exist, an error is thrown.
}
\examples{
# Delete a custom grid for the srgamma function
delete_build_in_grid(sampling_function = "srgamma", grid_type = "custom")

# Delete a scaled grid for the srnorm function
delete_build_in_grid(sampling_function = "srnorm", grid_type = "scaled")

}
