% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_optimizer.R
\name{find_optimal_grid}
\alias{find_optimal_grid}
\title{}
\usage{
find_optimal_grid(
  dendata,
  density_name,
  lb,
  rb,
  modes,
  f,
  h = NULL,
  h_prime = NULL,
  cdf = NULL,
  verbose,
  target_sample_size,
  theta,
  grid_limits
)
}
\description{

}
