% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_builder.R
\name{find_steps}
\alias{find_steps}
\title{Steps Builder}
\usage{
find_steps(lb = -Inf, rb = Inf, a, th, mode, mode_i, mode_n, f, h_prime, h)
}
\arguments{
\item{lb}{scaler density lower bound}

\item{rb}{scaler density upper bound}

\item{a}{scaler step area}

\item{th}{scaler pre_acceptance threshold}

\item{mode}{vector density modes}

\item{mode_i}{mode index}

\item{mode_n}{total number of modes}

\item{f}{density function}

\item{h_prime}{first derivative of h}

\item{h}{log transform of the density function}
}
\description{
Steps Builder
}
