% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_builder.R
\name{grid_builder}
\alias{grid_builder}
\title{Grid Builder}
\usage{
grid_builder(
  lb,
  rb,
  a,
  modes,
  f = NULL,
  h = NULL,
  h_prime = NULL,
  cdf = NULL,
  stps,
  lstpsp,
  rstpsp,
  theta,
  grid_range
)
}
\description{
Grid Builder
}
