% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_builder.R
\name{grid_prob_error_checking_and_preparation}
\alias{grid_prob_error_checking_and_preparation}
\title{check build_grid parameters validity}
\usage{
grid_prob_error_checking_and_preparation(gp_env)
}
\description{
check build_grid parameters validity
}
