% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_builder.R
\name{load_grid}
\alias{load_grid}
\title{Title}
\usage{
load_grid(grid_name)
}
\arguments{
\item{grid_name}{the name of the stored grid that has been stored using save_grid()}
}
\value{
return a saved grid from user data directory
}
\description{
Title
}
