% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_info.R
\name{print.grid}
\alias{print.grid}
\title{Print Method for Grid Objects}
\usage{
\method{print}{grid}(x, ...)
}
\arguments{
\item{x}{A list generated using STORS' `build_grid()` or `grid_optimizer()` functions.}

\item{...}{Additional arguments passed to the `print` function.}
}
\value{
Prints a summary of the grid's properties, but does not return any value.
}
\description{
The function displays detailed information about the grid object created by STORS' `build_grid()` or `grid_optimizer()` functions.
 This includes the number of steps within the grid, the range of values covered by the grid, and the grid's sampling efficiency.
  This information is crucial for understanding the structure and performance of the grid in sampling processes.
}
\details{
This method extends the generic `print` function for objects of class `grid`.
It prints the provided grid's features such as the number of steps, steps limit, and efficiency.
}
\examples{
# Define the density function, its logarithm,
#and its derivative for the standard normal distribution
modes_norm = 0
f_norm <- function(x) { 1 / sqrt(2 * pi) * exp(-0.5 * x^2) }
h_norm <- function(x) { log(f_norm(x)) }
h_prime_norm <- function(x) { -x }

# Build a dense grid for the standard normal distribution
norm_grid = build_grid(lb = -Inf, rb = Inf, mode = modes_norm,
 f = f_norm, h = h_norm, h_prime = h_prime_norm, steps = 1000)

# Print the properties of the generated grid
print(norm_grid)

}
