% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_builder.R
\name{print_grids}
\alias{print_grids}
\title{Print All Grids}
\usage{
print_grids()
}
\description{
Prints details of all grids stored by the user, including grid name, size, efficiency, etc.
}
\examples{
# To print details of all stored grids
print_grids()
}
