% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_builder.R
\name{save_grid}
\alias{save_grid}
\title{function to save grids that created by the user}
\usage{
save_grid(grid, file_name)
}
\arguments{
\item{grid}{an optimize grid generated using build_grid() function}

\item{file_name}{the name of the file that is holding the grid in the used data directory}
}
\description{
function to save grids that created by the user
}
