% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_builder.R
\name{set_grid}
\alias{set_grid}
\title{Grid builder}
\usage{
set_grid(
  lb = -Inf,
  rb = Inf,
  mode,
  f = f_norm,
  h = h_norm,
  h_prime = h_prime_norm
)
}
\arguments{
\item{lb}{scaler density lower bound}

\item{rb}{scaler density upper bound}

\item{mode}{vector density modes}

\item{f}{density function}

\item{h}{log transform of the density function}

\item{h_prime}{first derivative of h}
}
\value{
message
}
\description{
Grid builder
}
