% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stors.R
\name{snorm}
\alias{snorm}
\title{stros for sampling}
\usage{
snorm(n, mean = 0, sd = 1)
}
\arguments{
\item{n}{sample size}

\item{grid}{proposal grid}

\item{f}{target density}
}
\value{
sample of size n
}
\description{
stros for sampling
}
