% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srchisq.R
\name{srchisq}
\alias{srchisq}
\alias{srchisq_truncate}
\title{Sampling from Normal Distribution}
\usage{
srchisq(n)

srchisq_truncate(xl = 0, xr = Inf)
}
\arguments{
\item{n}{Integer sample size.}

\item{xl}{Lower bound for truncation.}

\item{xr}{Upper bound for truncation.}
}
\value{
\code{srchisq()} returns a sample of size \code{n} from a standard normal distribution.

\code{srchisq_truncate()} returns a function that, when called with a sample size \code{n}, generates \code{n} samples from a normal distribution truncated between \code{xl} and \code{xr}.
}
\description{
Sampling from the Normal distribution using Stors.
}
\details{
The function \code{srchisq()} is used for sampling from a standard Normal distribution (mean = 0, standard deviation = 1).



Normal distribution has the density:

\deqn{ f(x) = \frac{1}{\sigma\sqrt{2\pi}} e^{-\frac{(x - \mu)^2}{2\sigma^2}} }

Where \eqn{\mu} is the mean and \eqn{\sigma} is the standard deviation.

\code{srchisq_truncate()}, this function allows sampling from a standard normal distribution that is truncated within specified bounds.
 It is particularly useful when the area of interest in a normal distribution is limited to a specific range.
 The function first validates the truncation bounds to ensure they are within the allowable range of the distribution and then creates a tailored sampling function based on these bounds.
}
\examples{
# Generating Samples from a Standard Normal Distribution
# This example illustrates how to generate 10 samples from a standard normal distribution.
# It first optimizes the grid for sampling using \code{grid_optimizer},
# and then generates samples using \code{srchisq}.

# Optimize the grid for the standard normal distribution
grid_optimizer("srchisq")

# Generate and print 10 samples from the standard normal distribution
samples <- srchisq(10)
print(samples)


# Generating Samples from a Truncated Standard Normal Distribution
# This example demonstrates how to generate 100,
# samples from a standard normal distribution truncated in the range [-2, 2].

# Create the truncated sampling function
norm_trunc <- srchisq_truncate(xl = -2, xr = 2)

# Generate 100 samples
sample <- norm_trunc(100)

# Plot a histogram of the samples
hist(sample, main = "Histogram of Truncated Normal Samples", xlab = "Value", breaks = 20)

}
