% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srgamma.R
\name{srgamma}
\alias{srgamma}
\title{Sampling from Normal Distribution}
\usage{
srgamma(n = 1, x = NULL)
}
\arguments{
\item{n}{Integer sample size.}
}
\value{
\code{srgamma()} returns a sample of size \code{n} from a standard normal distribution.
}
\description{
Sampling from the Normal distribution using Stors.
}
\details{
The function \code{srgamma()} is used for sampling from a standard Normal distribution (mean = 0, standard deviation = 1).



Normal distribution has the density:

\deqn{ f(x) = \frac{1}{\sigma\sqrt{2\pi}} e^{-\frac{(x - \mu)^2}{2\sigma^2}} }

Where \eqn{\mu} is the mean and \eqn{\sigma} is the standard deviation.
}
\examples{
# Generating Samples from a Standard Normal Distribution
# This example illustrates how to generate 10 samples from a standard normal distribution.
# It first optimizes the grid for sampling using \code{grid_optimizer},
# and then generates samples using \code{srgamma}.

# Optimize the grid for the standard normal distribution
grid_optimizer("srgamma")

# Generate and print 10 samples from the standard normal distribution
samples <- srgamma(10)
print(samples)

}
