% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srgamma.R
\name{srgamma_optimize}
\alias{srgamma_optimize}
\title{Optimizing Gamma Distribution Grid}
\usage{
srgamma_optimize(
  shape = NULL,
  rate = NULL,
  xl = NULL,
  xr = NULL,
  scale = NULL,
  steps = 4091,
  grid_range = NULL,
  theta = NULL,
  target_sample_size = 1000,
  verbose = FALSE
)
}
\arguments{
\item{shape}{(optional) Numeric. The shape parameter (\eqn{\alpha}) of the Gamma distribution. Defaults to \code{1}.}

\item{rate}{(optional) Numeric. The rate parameter (\eqn{\beta}) of the Gamma distribution. Defaults to \code{1}.}

\item{xl}{Numeric. Left truncation bound for the target distribution. Defaults to \code{0}, representing no left truncation.}

\item{xr}{Numeric. Right truncation bound for the target distribution. Defaults to \code{Inf}, representing no right truncation.}

\item{scale}{(optional) Numeric. The scale parameter of the Gamma distribution. Defaults to \code{1}.}

\item{steps}{(optional) Integer. Desired number of steps in the proposal grid. Defaults to \code{4091}.}

\item{grid_range}{(optional) Numeric vector. Specifies the range for optimizing the steps part of the proposal grid. Defaults to \code{NULL}, indicating automatic range selection.}

\item{theta}{(optional) Numeric. A parameter for grid optimization. Defaults to \code{NULL}.}

\item{target_sample_size}{(optional) Integer. Target sample size for grid optimization. Defaults to \code{1000}.}

\item{verbose}{Boolean. If \code{TRUE}, detailed optimization information, including areas and steps, will be displayed. Defaults to \code{FALSE}.}
}
\value{
A list containing the optimized grid and related parameters for the specified Gamma distribution. The grid is also cached for internal use.
\describe{
  \item{\code{grid_data}}{Detailed information about the grid steps, including \code{x}, \code{s_upper}, \code{p_a}, and \code{s_upper_lower}.}
  \item{\code{areas}}{The areas under the left tail, steps, and right tail of the proposal distribution.}
  \item{\code{steps_number}}{The number of steps in the proposal grid.}
  \item{\code{f_params}}{The parameters (\code{shape} and \code{rate}) of the Gamma distribution.}
}
}
\description{
The \code{srgamma_optimize()} function generates an optimized proposal grid for a targeted Gamma distribution.
The grid can be customized and adjusted based on various options provided by the user.
}
\details{
When \code{srgamma_optimize()} is explicitly called:
\itemize{
  \item A grid is created and cached. If no parameters are provided, a standard grid is created with \code{shape = 1} and \code{rate = 1}.
  \item Providing \code{shape} and \code{rate} creates a custom grid, which is cached for use with \code{srgamma_custom()}.
  \item The optimization process can be controlled via parameters such as \code{steps}, \code{grid_range}, or
  \code{theta}. If no parameters are provided, the grid is optimized via brute force based on the
  \code{target_sample_size}.
}
}
\examples{
# Generate a standard grid with shape = 1 and rate = 1
standard_grid <- srgamma_optimize()

# Generate a custom grid with shape = 2 and rate = 3
custom_grid <- srgamma_optimize(shape = 2, rate = 3)

}
\seealso{
\code{\link{srgamma_custom}}: Function to sample from a custom grid generated by \code{srgamma_optimize()}.
}
