% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srlaplace.R
\name{srlaplace_optimize}
\alias{srlaplace_optimize}
\title{Optimizing Laplace Distribution Grid}
\usage{
srlaplace_optimize(
  mu = NULL,
  b = NULL,
  xl = NULL,
  xr = NULL,
  steps = 4091,
  grid_range = NULL,
  theta = NULL,
  target_sample_size = 1000,
  verbose = FALSE,
  symmetric = FALSE
)
}
\arguments{
\item{mu}{(optional) Numeric, location parameter.}

\item{b}{(optional) Numeric, scale parameter.}

\item{xl}{Numeric. Left truncation bound for the target distribution. Defaults to \code{-Inf}, representing no left truncation.}

\item{xr}{Numeric. Right truncation bound for the target distribution. Defaults to \code{Inf}, representing no right truncation.}

\item{steps}{(optional) Integer. Desired number of steps in the proposal grid. Defaults to \code{NULL}, which means the number of steps is determined automatically during optimization.}

\item{grid_range}{(optional) Numeric vector. Specifies the range for optimizing the steps part of the proposal grid. Defaults to \code{NULL}, indicating automatic range selection.}

\item{theta}{(optional) Numeric. A parameter for grid optimization. Defaults to \code{NULL}.}

\item{target_sample_size}{(optional) Integer. Target sample size for grid optimization. Defaults to \code{1000}.}

\item{verbose}{Boolean. If \code{TRUE}, detailed optimization information, including areas and steps, will be displayed. Defaults to \code{FALSE}.}

\item{symmetric}{Boolean. If \code{TRUE}, the proposal will target only the right tail of the distribution, reducing the size of the cached grid and making sampling more memory-efficient.
An additional uniform random number will be sampled to determine the sample's position relative to the mode of the distribution.
While this improves memory efficiency, the extra sampling may slightly impact performance, especially when the grid efficiency is close to 1. Defaults to \code{FALSE}.}
}
\value{
The user does not need to store the returned value, because the package internally cashes the grid. However, we explain here the full returned grid for advanced users.

A list containing the optimized grid and related parameters for the specified built-in distribution:
\describe{
\item{\code{grid_data}}{A data frame with detailed information about the grid steps, including:
\describe{
\item{\code{x}}{The start point of each step on the x-axis.}
\item{\code{s_upper}}{The height of each step on the y-axis.}
\item{\code{p_a}}{Pre-acceptance probability for each step.}
\item{\code{s_upper_lower}}{A vector used to scale the uniform random number when the sample is accepted.}
}}
\item{\code{areas}}{A numeric vector containing the areas under:
\describe{
\item{\code{left_tail}}{The left tail bound.}
\item{\code{steps}}{The middle steps.}
\item{\code{right_tail}}{The right tail bound.}
}}
\item{\code{steps_number}}{An integer specifying the number of steps in the proposal.}
\item{\code{sampling_probabilities}}{A numeric vector with:
\describe{
\item{\code{left_tail}}{The probability of sampling from the left tail.}
\item{\code{left_and_middle}}{The combined probability of sampling from the left tail and middle steps.}
}}
\item{\code{unif_scaler}}{A numeric scalar, the inverse probability of sampling from the steps part of the proposal (\eqn{\frac{1}{p(lb < x < rb)}}). Used for scaling uniform random values.}
\item{\code{lt_properties}}{A numeric vector of 5 values required for Adaptive Rejection Sampling (ARS) in the left tail.}
\item{\code{rt_properties}}{A numeric vector of 6 values required for ARS in the right tail.}
\item{\code{alpha}}{A numeric scalar representing the uniform step area.}
\item{\code{tails_method}}{A string, either \code{"ARS"} (Adaptive Rejection Sampling) or \code{"IT"} (Inverse Transform), indicating the sampling method for the tails.}
\item{\code{grid_bounds}}{A numeric vector specifying the left and right bounds of the target density.}
\item{\code{cnum}}{An integer representing the cache number of the created grid in memory.}
\item{\code{symmetric}}{A numeric scalar indicating the symmetry point of the grid, or \code{NULL} if not symmetric.}
\item{\code{f_params}}{A list of parameters for the target density that the proposal grid is designed for.}
\describe{
\item{\code{mu}}{is the location parameter (location of the distribution).}
\item{\code{b}}{is the scale parameter, which controls the spread of the distribution (\code{b > 0}).}
}
\item{\code{is_symmetric}}{A logical value indicating whether the proposal grid is symmetric.}
\item{\code{grid_type}}{A string indicating the type of the generated grid:
\describe{
\item{\code{"scaled"}}{The grid is "scalable" and standardized with \code{mu = 0} and \code{b = 1}. This is used when parameters \code{mu} and \code{b} are either \code{NULL} or not provided. Scalable grids are compatible with \code{\link{srlaplace}}.}
\item{\code{"custom"}}{The grid is "custom" when either \code{mu} or \code{b} is provided. Custom grids are compatible with \code{\link{srlaplace_custom}}.}
}}
\item{\code{target_function_area}}{A numeric scalar estimating the area of the target distribution.}
\item{\code{dens_func}}{A string containing the hardcoded density function.}
\item{\code{density_name}}{A string specifying the name of the target density distribution.}
\item{\code{lock}}{An identifier used for saving and loading the grid from disk.}
}
}
\description{
The \code{srlaplace_optimize()} function generates an optimized proposal grid for a targeted Laplace Distribution.
The grid can be customized and adjusted based on various options provided by the user.
}
\details{
When \code{srlaplace_optimize()} is explicitly called:
\itemize{
\item A grid is created and cached. If no parameters are provided, a standard grid is created (\code{mu = 0}, \code{b = 1}).
\item Providing \code{mu} or \code{b} creates a custom grid, which is cached for use with \code{srlaplace_custom()}.
\item The optimization process can be controlled via parameters such as \code{steps}, \code{grid_range}, or
\code{theta}. If no parameters are provided, the grid is optimized via brute force based on the.
\code{target_sample_size}.
}
}
\examples{
# Generate scalable grid that with mu = 0 and b = 1, that has 4096 steps
scalable_grid <- srlaplace_optimize(steps = 4096)

# Generate custom grid that with mu = 2 and b = 1
scalable_grid <- srlaplace_optimize(mu = 2, b = 1)

}
\seealso{
\code{\link{srlaplace}}: Function to sample from a scalable grid generated by \code{srlaplace_optimize()}.
\code{\link{srlaplace_custom}}: Function to sample from a custom grid tailored to user specifications.
}
