% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srnorm.R
\name{srnorm_optimize}
\alias{srnorm_optimize}
\title{Optimizing Normal Distribution Grid}
\usage{
srnorm_optimize(
  mean = NULL,
  sd = NULL,
  xl = -Inf,
  xr = Inf,
  steps = NULL,
  grid_range = NULL,
  theta = NULL,
  target_sample_size = 1000,
  verbose = FALSE,
  symmetric = NULL
)
}
\arguments{
\item{mean}{Numeric. Mean parameter of the Normal distribution.}

\item{sd}{Numeric. Standard deviation of the target Normal distribution.}

\item{xl}{(optional) Scalar lower bounds.}

\item{xr}{(optional) Scalar upper bounds.}

\item{steps}{(optional) Scalar integer indicating the number of steps in the proposal distribution.}

\item{grid_range}{(optional) Optional Vector of two elements specifying the start and end points for constructing steps along the x-axis.}

\item{theta}{(optional) the minimum acceptable pre-acceptance value for the grid which indirectly determine the number of steps. Can not be specified at the same time as `steps`.}

\item{target_sample_size}{(optional) Scalar integer indicating the target sample size optimizer will simulate batches of this size when benchmarking.}

\item{verbose}{(optional) Logical if set to `TRUE`, a table detailing the optimization areas and steps will be displayed during grid optimization.}

\item{symmetric}{(optional) Logical if True then only one half of the grid is built ti the left to the mode and a coin toss is used to determine weather to reflect around the mode, this allows a dencer grid for the same memory at the cost of an additional uniform draw "default is `FALSE`"}
}
\description{
Optimizing Normal Distribution Grid
}
