% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stors.R
\name{stors}
\alias{stors}
\title{Sampling Function for Users' Grid}
\usage{
stors(grid)
}
\arguments{
\item{grid}{The sampling grid created by the user.}
}
\value{
Returns a sampling function that can be used to generate samples from the input `grid`.
}
\description{
This function generates a sampling function based on a grid created and optimized by the user using the `build_grid()` function. The resulting sampling function can then be used to produce samples.
}
