% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stors.R
\name{stors3}
\alias{stors3}
\title{Title}
\usage{
stors3(n, x, lower, pa, steps, pro, unis, upper, lt, rt, f)
}
\arguments{
\item{n}{}

\item{x}{}

\item{lower}{}

\item{pa}{}

\item{steps}{}

\item{pro}{}

\item{unis}{}

\item{upper}{}

\item{lt}{}

\item{rt}{}

\item{f}{}
}
\description{
Title
}
