% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace.R
\name{trunclaplace}
\alias{trunclaplace}
\title{I will stick with this function until I found a proper solution.
Title}
\usage{
trunclaplace(xl, xr)
}
\arguments{
\item{n}{}

\item{l}{}

\item{r}{}
}
\description{
I will stick with this function until I found a proper solution.
Title
}
