% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_builder.R
\name{delete_grid}
\alias{delete_grid}
\title{Delete Stored Grid}
\usage{
delete_grid(grid_name)
}
\arguments{
\item{grid_name}{string specifying the name of the grid to be deleted.}
}
\value{
Shows an error if `grid_name` does not exist. Otherwise, it successfully deletes the grid without returning any value upon completion.
}
\description{
Deletes a grid previously stored by the user using the `save_grid()` function.
}
