% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_builder.R
\name{grid_builder}
\alias{grid_builder}
\title{Grid Builder}
\usage{
grid_builder(lb = -Inf, rb = Inf, a, th, mode, f, h, h_prime)
}
\arguments{
\item{lb}{scaler density lower bound}

\item{rb}{scaler density upper bound}

\item{a}{scaler step area}

\item{th}{scaler pre_acceptance threshold}

\item{mode}{vector density modes}

\item{f}{density function}

\item{h}{log transform of the density function}

\item{h_prime}{first derivative of h}
}
\value{
list including proposal distribution properties
}
\description{
Grid Builder
}
