% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_optimizer.R
\name{grid_optimizer}
\alias{grid_optimizer}
\title{Optimize Built-in Grid}
\usage{
grid_optimizer(
  density_name = stors_env$grids$biultin$names,
  verbose = FALSE,
  target_sample_size = 1000,
  steps = NA,
  grid_range,
  theta = 0
)
}
\arguments{
\item{density_name}{string specifying the name of the distribution.}
}
\value{
This function generates and stores an optimized grid for the specified built-in distribution in R's internal data directory.
}
\description{
This function optimizes and stores the sampling grid for STROS built-in distributions.
}
\details{
Before users can sample from a built-in distribution in the STROS package, they must first optimize the sampling grid using this function.
}
\examples{
# optimize grid to sample 10 values from normal distribution N(0,1) 
library(stors)
grid_optimizer("srnorm")
srnorm(10)

}
\seealso{
[grid_optimizer()] is used to optimize and cache grid to sample from multible distrubution such as [srnorm()] and [srgamma()]
}
