% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_info.R
\name{load_grid}
\alias{load_grid}
\title{Load Stored Grid}
\usage{
load_grid(grid_name)
}
\arguments{
\item{grid_name}{A string specifying the name of the grid to be loaded.}
}
\value{
Returns a list of the grid stored under `grid_name` in R's internal data directory. If the specified grid name does not exist, it shows an error.
}
\description{
This function loads a grid into memory that was previously saved using the `save_grid()` function.
}
