% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_info.R
\name{load_grid}
\alias{load_grid}
\title{Load Stored Grid}
\usage{
load_grid(grid_name)
}
\arguments{
\item{grid_name}{A string specifying the name of the grid to be loaded.}
}
\value{
Returns a list representing the grid stored under \code{grid_name} in R's internal data directory. If the grid corresponding to the specified name does not exist, an error message is displayed.
}
\description{
This function loads a grid into memory that was previously saved using the \code{save_grid()} function. It is useful for retrieving saved grids for further analysis or processing.
}
\examples{
# First, let's create a grid to sample from a standard normal distribution
f_normal <- function(x) { 0.3989423 * exp(-0.5 * x^2) }
normal_grid = build_grid(f = f_normal, modes = 0, lb = -Inf, rb = Inf, steps = 1000)
print(normal_grid)

# Then, save this grid in R's internal data directory using `save_grid()` with the name "normal"
save_grid(normal_grid, "normal")

# Now, in case the R session is restarted and the grid is no longer in memory,
# it can be loaded from the machine as follows:
loaded_normal_grid <- load_grid("normal")
print(loaded_normal_grid)
}
