% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_builder.R
\name{save_grid}
\alias{save_grid}
\title{Save User Grid}
\usage{
save_grid(grid, grid_name)
}
\arguments{
\item{grid}{list representing an optimized grid generated using the `build_grid()` function.}

\item{grid_name}{string specifying the name under which the grid will be saved.}
}
\value{
This function will produce an error if the grid is not generated by the `build_grid()` function. Otherwise, it successfully saves the grid without returning any value upon completion.
}
\description{
This function stores grids generated by the `build_grid()` function. It is useful when users want to reuse a grid across multiple R sessions.
}
