% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_info.R
\name{save_grid}
\alias{save_grid}
\title{Save User Grid}
\usage{
save_grid(grid, grid_name)
}
\arguments{
\item{grid}{list representing an optimized grid generated using the `build_grid()` function.}

\item{grid_name}{string specifying the name under which the grid will be saved.}
}
\value{
This function will produce an error if the grid is not generated by the `build_grid()` function. Otherwise, it successfully saves the grid without returning any value upon completion.
}
\description{
This function stores grids generated by the `build_grid()` function in R's internal data directory. It is useful when users want to reuse a grid across multiple R sessions.
}
\examples{
# First, let's create a grid to sample from a standard normal distribution
f_normal <- function(x) { 0.3989423 * exp(-0.5 * x^2) }
normal_grid = build_grid(f = f_normal, modes = 0, lb = -Inf, rb = Inf, steps = 1000)
print(normal_grid)

# Then, we can save this grid in R's internal data directory using `save_grid()`
# with the name "normal"
save_grid(normal_grid, "normal")

# To make sure the 'normal_grid' has been stored in R's internal data directory,
# we can print all saved grids using `print_grids()`
print_grids()

}
