% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srnorm.R
\name{srnorm}
\alias{srnorm}
\alias{srsnorm}
\title{Sampling from Normal distribution}
\usage{
srnorm(n)

srsnorm(n, mu = 0, sd = 1)
}
\arguments{
\item{n}{Integer sample size}

\item{mu}{Scalar mean}

\item{sd}{Scalar standard deviations}
}
\value{
sample of size n
}
\description{
Sampling from Normal distribution using stors.
}
\details{
The function srnorm() is used to sample from a standard Normal distribution with a mean equal to zero and a standard deviation equal to 1. In contrast, srsnorm() is used to sample from any Normal distribution by accepting the mean and standard deviation as inputs. The reason for having these two separate functions is to improve performance. The Stors algorithm is so fast that even simple operations like multiplication and addition could impact its speed.
}
