% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srbeta.R
\name{srbeta_optimize}
\alias{srbeta_optimize}
\title{Optimizing Beta Distribution Grid}
\usage{
srbeta_optimize(
  shape1 = 2,
  shape2 = 2,
  xl = 0,
  xr = 1,
  steps = NULL,
  grid_range = NULL,
  theta = NULL,
  target_sample_size = 1000,
  verbose = FALSE
)
}
\arguments{
\item{shape1}{(optional) Numeric. The first shape parameter (\eqn{\alpha}) of the Beta distribution. Defaults to \code{1}.}

\item{shape2}{(optional) Numeric. The second shape parameter (\eqn{\beta}) of the Beta distribution. Defaults to \code{1}.}

\item{xl}{Numeric. Left truncation bound for the target distribution. Defaults to \code{0}, as the Beta distribution is defined only on the interval \code{[0, 1]}.}

\item{xr}{Numeric. Right truncation bound for the target distribution. Defaults to \code{1}, as the Beta distribution is defined only on the interval \code{[0, 1]}.}

\item{steps}{(optional) Integer. Desired number of steps in the proposal grid. Defaults to \code{4091}.}

\item{grid_range}{(optional) Numeric vector. Specifies the range for optimizing the steps part of the proposal grid. Defaults to \code{NULL}, indicating automatic range selection.}

\item{theta}{(optional) Numeric. A parameter for grid optimization. Defaults to \code{NULL}.}

\item{target_sample_size}{(optional) Integer. Target sample size for grid optimization. Defaults to \code{1000}.}

\item{verbose}{Boolean. If \code{TRUE}, detailed optimization information, including areas and steps, will be displayed. Defaults to \code{FALSE}.}
}
\value{
A list containing the optimized grid and related parameters for the specified Beta distribution. The grid is also cached for internal use.
\describe{
\item{\code{grid_data}}{Detailed information about the grid steps, including \code{x}, \code{s_upper}, \code{p_a}, and \code{s_upper_lower}.}
\item{\code{areas}}{The areas under the left tail, steps, and right tail of the proposal distribution.}
\item{\code{steps_number}}{The number of steps in the proposal grid.}
\item{\code{f_params}}{The parameters (\code{shape1} and \code{shape2}) of the Beta distribution.}
}
}
\description{
The \code{srbeta_optimize()} function generates an optimized proposal grid for a targeted Beta distribution.
The grid can be customized and adjusted based on various options provided by the user.
}
\details{
When \code{srbeta_optimize()} is explicitly called:
\itemize{
\item A grid is created and cached. If no parameters are provided, a standard grid is created with \code{shape1 = 1} and \code{shape2 = 1}.
\item Providing \code{shape1} and \code{shape2} creates a custom grid, which is cached for use with \code{srbeta_custom()}.
\item The optimization process can be controlled via parameters such as \code{steps}, \code{grid_range}, or
\code{theta}. If no parameters are provided, the grid is optimized via brute force based on the
\code{target_sample_size}.
}
}
\examples{
# Generate a standard grid with shape1 = 1 and shape2 = 1
standard_grid <- srbeta_optimize()

# Generate a custom grid with shape1 = 2 and shape2 = 3
custom_grid <- srbeta_optimize(shape1 = 2, shape2 = 3)

}
\seealso{
\code{\link{srbeta_custom}}: Function to sample from a custom grid generated by \code{srbeta_optimize()}.
}
