% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srexp.R
\name{srexp}
\alias{srexp}
\alias{srexp_scaled}
\alias{srexp_truncate}
\title{Sampling from Exponential Distribution}
\usage{
srexp(n)

srexp_scaled(n, scale = 1)

srexp_truncate(xl, xr)
}
\arguments{
\item{n}{Integer sample size.}

\item{scale}{Scalar scale parameter.}

\item{xl}{Lower bound for truncation.}

\item{xr}{Upper bound for truncation.}
}
\value{
\code{srexp()} returns a sample of size \code{n} from a standard exponential distribution.

\code{srexp_scaled()} returns a sample of size \code{n} from an exponential distribution with the specified scale parameter.

\code{srexp_truncate()} returns a function that generates \code{n} samples from a truncated exponential distribution between \code{xl} and \code{xr}.
}
\description{
Sampling from the exponential distribution using Stors.

Sampling from any exponential distribution by specifying the scale parameter.

Creating a sampling function for a truncated exponential distribution.
}
\details{
The function \code{srexp()} is used for sampling from a standard exponential distribution with a rate parameter of 1.

Exponential distribution has the density:
\deqn{ f(x; \lambda) = \lambda e^{-\lambda x} }
Where \eqn{\lambda} is the rate parameter.

\code{srexp_truncate()} is used for sampling from a standard exponential distribution truncated within specified bounds.
The function validates the truncation bounds before creating the sampling function.
}
\examples{

# Optimize the grid for Exponential distribution
grid_optimizer("srexp")

# Generating Samples from a Standard Exponential Distribution
# This example illustrates how to generate 10 samples,
# from a standard exponential distribution.
samples <- srexp(10)
print(samples)

# Generating Samples from an Exponential Distribution with a Specific Scale
# This example demonstrates how to generate 10 samples,
# from an exponential distribution with a scale parameter of 2.
samples <- srexp_scaled(n = 10, scale = 2)
print(samples)

# Generating Samples from a Truncated Exponential Distribution
# This example demonstrates how to generate 100 samples,
# from an exponential distribution truncated in the range [1, 3].
exp_trunc <- srexp_truncate(xl = 1, xr = 3)
samples <- exp_trunc(100)
hist(samples, main = "Histogram of Truncated Exponential Samples", xlab = "Value", breaks = 20)

}
