% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srexp.R
\name{srexp}
\alias{srexp}
\alias{srexp_custom}
\title{Sampling from Exponential Distribution}
\usage{
srexp(n = 1, rate = 1, x = NULL)

srexp_custom(n = 1, x = NULL)
}
\arguments{
\item{n}{Integer sample size.}
}
\value{
\code{srexp()} returns a sample of size \code{n} from a standard exponential distribution.
}
\description{
Sampling from the exponential distribution using Stors.
}
\details{
The function \code{srexp()} is used for sampling from a standard exponential distribution with a rate parameter of 1.

Exponential distribution has the density:
\deqn{ f(x; \lambda) = \lambda e^{-\lambda x} }
Where \eqn{\lambda} is the rate parameter.
}
\examples{

# Optimize the grid for Exponential distribution
grid_optimizer("srexp")

# Generating Samples from a Standard Exponential Distribution
# This example illustrates how to generate 10 samples,
# from a standard exponential distribution.
samples <- srexp(10)
print(samples)

}
