% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srlaplace.R
\name{srlaplace}
\alias{srlaplace}
\alias{srlaplace_custom}
\title{Sampling from Laplace Distribution}
\usage{
srlaplace(n = 1, mu = 0, b = 1, x = NULL)

srlaplace_custom(n = 1, x = NULL)
}
\arguments{
\item{n}{Integer sample size.}

\item{mu}{Scalar location parameter.}

\item{b}{Scalar scale parameter.}
}
\value{
\code{srlaplace()} returns a sample of size \code{n} from a standard Laplace distribution.

\code{srlaplace_scaled()} returns a sample of size \code{n} from a Laplace distribution with location \eqn{\mu} and scale \eqn{b}.
}
\description{
Sampling from the Laplace distribution using Stors.

Sampling from any Laplace distribution specifying the location and scale.
}
\details{
The function \code{srlaplace()} is used for sampling from a standard Laplace distribution (location = 0, scale = 1).

Laplace distribution has the density:
\deqn{ f(x | \mu, b) = \frac{1}{2b} \exp\left(-\frac{|x - \mu|}{b}\right) }
Where \eqn{\mu} is the location parameter and \eqn{b} is the scale parameter.
}
\examples{
# Optimize the grid for the Laplace distribution
grid_optimizer("srlaplace")
# Generating Samples from a Standard Laplace Distribution
# This example illustrates how to generate 10 samples from a standard Laplace distribution.
samples <- srlaplace(10)
print(samples)


# This example demonstrates how to generate 10,
# Generating Samples from a Laplace Distribution with Specific Location and Scale
# samples from a Laplace distribution with location = 2 and scale = 3.
samples <- srlaplace_scaled(n = 10, mu = 2, b = 3)
print(samples)

}
