% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_srnorm.R, R/srnorm.R, R/srnorm_symmetric.R
\name{old_srnorm}
\alias{old_srnorm}
\alias{srnorm}
\alias{srnorm_symmetric}
\alias{old_srnorm_scaled}
\alias{srnorm_scaled}
\alias{old_srnorm_truncate}
\alias{srnorm_truncate}
\title{Sampling from Normal Distribution}
\usage{
old_srnorm(n)

srnorm(n)

srnorm_symmetric(n)

old_srnorm_scaled(n, mu = 0, sd = 1)

srnorm_scaled(n, mu = 0, sd = 1)

old_srnorm_truncate(xl = -Inf, xr = Inf)

srnorm_truncate(xl = -Inf, xr = Inf)
}
\arguments{
\item{n}{Integer sample size.}

\item{mu}{Scalar mean.}

\item{sd}{Scalar standard deviation.}

\item{xl}{Lower bound for truncation.}

\item{xr}{Upper bound for truncation.}
}
\value{
\code{srnorm()} returns a sample of size \code{n} from a standard normal distribution.

\code{srnorm()} returns a sample of size \code{n} from a standard normal distribution.

\code{srnorm()} returns a sample of size \code{n} from a standard normal distribution.

\code{srnorm_scaled()} returns a sample of size \code{n} from a normal distribution with mean \eqn{\mu} and standard deviation \eqn{\sigma}.

\code{srnorm_scaled()} returns a sample of size \code{n} from a normal distribution with mean \eqn{\mu} and standard deviation \eqn{\sigma}.

\code{srnorm_truncate()} returns a function that, when called with a sample size \code{n}, generates \code{n} samples from a normal distribution truncated between \code{xl} and \code{xr}.

\code{srnorm_truncate()} returns a function that, when called with a sample size \code{n}, generates \code{n} samples from a normal distribution truncated between \code{xl} and \code{xr}.
}
\description{
Sampling from the Normal distribution using Stors.

Sampling from the Normal distribution using Stors.

Sampling from the Normal distribution using Stors.
}
\details{
The function \code{srnorm()} is used for sampling from a standard Normal distribution (mean = 0, standard deviation = 1).



Normal distribution has the density:

\deqn{ f(x) = \frac{1}{\sigma\sqrt{2\pi}} e^{-\frac{(x - \mu)^2}{2\sigma^2}} }

Where \eqn{\mu} is the mean and \eqn{\sigma} is the standard deviation.

The function \code{srnorm()} is used for sampling from a standard Normal distribution (mean = 0, standard deviation = 1).



Normal distribution has the density:

\deqn{ f(x) = \frac{1}{\sigma\sqrt{2\pi}} e^{-\frac{(x - \mu)^2}{2\sigma^2}} }

Where \eqn{\mu} is the mean and \eqn{\sigma} is the standard deviation.

The function \code{srnorm()} is used for sampling from a standard Normal distribution (mean = 0, standard deviation = 1).



Normal distribution has the density:

\deqn{ f(x) = \frac{1}{\sigma\sqrt{2\pi}} e^{-\frac{(x - \mu)^2}{2\sigma^2}} }

Where \eqn{\mu} is the mean and \eqn{\sigma} is the standard deviation.

\code{srnorm_scaled()} allows sampling from any Normal distribution by specifying the mean and standard deviation.
The separation of these functions enhances performance, as the Stors algorithm is highly efficient, and even simple arithmetic can impact its speed.

\code{srnorm_scaled()} allows sampling from any Normal distribution by specifying the mean and standard deviation.
The separation of these functions enhances performance, as the Stors algorithm is highly efficient, and even simple arithmetic can impact its speed.

\code{srnorm_truncate()}, this function allows sampling from a standard normal distribution that is truncated within specified bounds.
 It is particularly useful when the area of interest in a normal distribution is limited to a specific range.
 The function first validates the truncation bounds to ensure they are within the allowable range of the distribution and then creates a tailored sampling function based on these bounds.

\code{srnorm_truncate()}, this function allows sampling from a standard normal distribution that is truncated within specified bounds.
 It is particularly useful when the area of interest in a normal distribution is limited to a specific range.
 The function first validates the truncation bounds to ensure they are within the allowable range of the distribution and then creates a tailored sampling function based on these bounds.
}
\examples{
# Generating Samples from a Standard Normal Distribution
# This example illustrates how to generate 10 samples from a standard normal distribution.
# It first optimizes the grid for sampling using \code{grid_optimizer},
# and then generates samples using \code{srnorm}.

# Optimize the grid for the standard normal distribution
grid_optimizer("srnorm")

# Generate and print 10 samples from the standard normal distribution
samples <- srnorm(10)
print(samples)


# Generating Samples from a Standard Normal Distribution
# This example illustrates how to generate 10 samples from a standard normal distribution.
# It first optimizes the grid for sampling using \code{grid_optimizer},
# and then generates samples using \code{srnorm}.

# Optimize the grid for the standard normal distribution
grid_optimizer("srnorm")

# Generate and print 10 samples from the standard normal distribution
samples <- srnorm(10)
print(samples)


# Generating Samples from a Standard Normal Distribution
# This example illustrates how to generate 10 samples from a standard normal distribution.
# It first optimizes the grid for sampling using \code{grid_optimizer},
# and then generates samples using \code{srnorm}.

# Optimize the grid for the standard normal distribution
grid_optimizer("srnorm")

# Generate and print 10 samples from the standard normal distribution
samples <- srnorm(10)
print(samples)

# Generating Samples from a Normal Distribution with Specific Mean and Standard Deviation
# This example demonstrates how to generate 10,
# samples from a normal distribution with a mean of 4 and a standard deviation of 2.

samples <- srnorm_scaled(n = 10, mu = 4, sd = 2)
print(samples)

# Generating Samples from a Normal Distribution with Specific Mean and Standard Deviation
# This example demonstrates how to generate 10,
# samples from a normal distribution with a mean of 4 and a standard deviation of 2.

samples <- srnorm_scaled(n = 10, mu = 4, sd = 2)
print(samples)

# Generating Samples from a Truncated Standard Normal Distribution
# This example demonstrates how to generate 100,
# samples from a standard normal distribution truncated in the range [-2, 2].

# Create the truncated sampling function
norm_trunc <- srnorm_truncate(xl = -2, xr = 2)

# Generate 100 samples
sample <- norm_trunc(100)

# Plot a histogram of the samples
hist(sample, main = "Histogram of Truncated Normal Samples", xlab = "Value", breaks = 20)

# Generating Samples from a Truncated Standard Normal Distribution
# This example demonstrates how to generate 100,
# samples from a standard normal distribution truncated in the range [-2, 2].

# Create the truncated sampling function
norm_trunc <- srnorm_truncate(xl = -2, xr = 2)

# Generate 100 samples
sample <- norm_trunc(100)

# Plot a histogram of the samples
hist(sample, main = "Histogram of Truncated Normal Samples", xlab = "Value", breaks = 20)

}
