% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stors.R
\name{stors}
\alias{stors}
\title{stros for sampling}
\usage{
stors(grid)
}
\arguments{
\item{grid}{proposal grid}

\item{n}{sample size}

\item{f}{target density}
}
\value{
sample of size n
}
\description{
stros for sampling
}
